### 2021-09-28 #3775 @whedon
[PRE REVIEW]: The AutoActive Research Environment
**Submitting author:** @kasbot (<a href="http://orcid.org/0000-0002-9983-0041">Kasper Bøtker Rasmussen</a>)
**Repository:** <a href="https://github.com/SINTEF/AutoActive-ActivityPresenter.git" target ="_blank">https://github.com/SINTEF/AutoActive-ActivityPresenter.git</a>
**Version:** v1.1.1
