### 2021-09-28 #3775 @danielskatz
👋 @kasbot - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Please feel free to use the command `@whedon generate pdf` after making changes to the .md file to make a new PDF. Whedon commands need to be the first entry in a new comment.
