### 2021-10-18 35677f1
checks session sturtures on save click
