### 2021-10-20 f47bce0
fixed bug when checking session structure
