### 2021-10-26 491a3e5
fixed bug when moving items inside tree
