### 2021-10-26 f697016
made it impossible to also move datapoints in savingtree if parent folder is not a temporary folder
