### 2021-11-16 #3775 @prashjha
Hello @kellerb, any chance you could review this submission? More details of this software are in this link: https://github.com/SINTEF/AutoActive-ActivityPresenter
