### 2022-01-13 #3775 @prashjha
Howdy, @erik-whiting! Would you be interested in reviewing this submission for JOSS?
