### 2022-01-13 #4061 @whedon
[REVIEW]: The AutoActive Research Environment
**Submitting author:** @kasbot (<a href="http://orcid.org/0000-0002-9983-0041">Kasper Bøtker Rasmussen</a>)
**Repository:** https://github.com/SINTEF/AutoActive-ActivityPresenter.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1

- [x] #4095
