### 2022-01-14 #4061 @prashjha
@erik-whiting, I will suggest opening an issue to rename the file as just LICENSE as it can get confusing.
