### 2022-01-15 SINTEF/AutoActive-ActivityPresenter#3 @erik-whiting
JOSS Review - Software Paper Review
Hi, I'm Erik Whiting, I'm one of the reviewers for your submission to the Journal of Open Source Software (JOSS). The review is being coordinated [here](https://github.com/openjournals/joss-reviews/issues/4061).

This issue is to note my review of the paper portion of your submission. For your convenience, I have copy and pasted the criteria set forth by the JOSS editors and my notes on why I think your paper does or does not meet the requirements. Please note, this is for the paper review portion only (there is also "General Checks," "Functionality," and "Documentation" reviews which I will open separate issues.

The **bolded** text is the JOSS-provided criteria for each item, and the text under it is my notes as they pertain to your submission:

**Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?**

Yes! I believe the "summary" section is well written and that non-specialist readers will comprehend what the software is intended to do.

**A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?**

Although the problem is well-stated in the "Statement of Need," you don't mention the target audience of your software. I can probably surmise that it's intended for biomechanical researchers, but I think the paper would benefit from a more explicit mention of  who this software is for specifically.

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**

You do mention 2 other open-source software packages in your paper, but I think you left a few things out. For example, I was able to find [this technology](https://www.pasco.com/products/software/capstone) that I think is similar to yours (though not exactly the same. [Link to video demonstration](https://www.pasco.com/resources/video/BmDeTdNIYSc)) as well as [this technology](https://sensiml.com/documentation/index.php) which seems very close to what your software does (synchronizing sensor data feeds with video feeds). These technologies are not open-source, but I think they (or similar technologies) are worth mentioning at least in general. I'm not saying you have to cite every sensor/video syncing software ever created, but I am saying that it's worth mentioning that similar technology does exist already, at least commercially.

It's possible I'm misunderstanding both the referenced technologies and yours so please tell me if this is the case.

**Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?**

The writing quality is excellent!

**References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?**

References are good :+1:
