### 2022-01-15 SINTEF/AutoActive-ActivityPresenter#4 @erik-whiting
JOSS Review - General Checks Review
Hi, I'm Erik Whiting, I'm one of the reviewers for your submission to the Journal of Open Source Software (JOSS). The review is being coordinated [here](https://github.com/openjournals/joss-reviews/issues/4061).

This issue is to note my review of the general checks portion of your submission. For your convenience, I have copy and pasted the criteria set forth by the JOSS editors and my notes on why I think your paper does or does not meet the requirements. Please note, this is for the general checks review portion only (there is also "Software Paper," "Functionality," and "Documentation" reviews which I will open separate issues.

The bolded text is the JOSS-provided criteria for each item, and the text under it is my notes as they pertain to your submission:

**Repository: Is the source code for this software available at the repository url?**

Yes

**License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?**

You have the appropriate license, but your file is called `LICENSE-2.0.txt`. Can you please rename this to just `LICENSE`?

**Contribution and authorship: Has the submitting author (@kasbot) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?**

Yes

**Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines**

Yes
