### 2022-01-15 #4061 @erik-whiting
@prashjha another question for you: Regarding the documentation checks, there's a checkbox item that says
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Most of the core functionality is well documented and It doesn't seem like there's an API for this, but there are MATLAB and Python libraries that are supposed to work with this submission's AAZ file format. These libraries seem like they're a separate from the submission, but the Python library doesn't appear to have great documentation. Is this part of "API method documentation" and should I advise the authors to provide some documentation for this (and possibly the MATLAB library too) or are these things not technically part of the submission and the existing documentation is good enough?
