### 2022-01-19 #4061 @prashjha
@erik-whiting, sorry I missed it. I am not 100% sure about this. If you think those scripts in Matlab/python are something users will need to use at some point in testing/using this library then the scripts should be well documented. I would consider the example or demo files integral to the library as these are the files any user needs to run first. If such files are too many, we of course can expect good documentation for all of them.
