### 2022-01-19 #4061 @erik-whiting
> @prashjha , Should we do any of these changes now? or should we wait untill both reviewers are completly finished?

@kasbot I'm not the editor but I think I can answer this for you. usually, once the issues are addressed, we just mark those check boxes as done. it's up to you when you want to start addressing issues but you don't have to wait for the reviews to finish.
