### 2022-01-23 SINTEF/AutoActive-ActivityPresenter#5 @AustinTSchaffer
Installation Steps and OS-Compatibility
Hello! JOSS reviewer here. Just wanted to make a quick comment about installation. I'm glad to see that this project is already published on the Microsoft Store. I recommend adding some notes to the Readme so that the word "install" shows up near the Microsoft Store link, for those quickly looking installation information.

Also, does this project support Linux/Mac via .NET Core, or is it 100% Windows-only? It might be good to mention OS-compatibility in the Readme as well.

Related to: https://github.com/openjournals/joss-reviews/issues/4061
