### 2022-01-23 SINTEF/AutoActive-ActivityPresenter#5 @AustinTSchaffer
I just noticed some notes at the very bottom of the Readme that mention OS-compatibility. It might be worth it to combine the install link and the notes about Xamarin into a single section titled "Installation", for clarity
