### 2022-01-23 SINTEF/AutoActive-ActivityPresenter#6 @AustinTSchaffer
Community Guidelines
>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I'm not seeing any references to contributions, issues, nor support in this repository, apart from what comes standard in the Apache 2.0 license. Could you please add a section that mentions these 3 things? It could be as simple as just mentioning that this project has an Apache 2.0 license and that people may submit issues/PRs on this GitHub repository.

Relevant thread: https://github.com/openjournals/joss-reviews/issues/4061
