### 2022-01-23 #4061 @AustinTSchaffer
@kasbot I left 2 issue threads on the repository about clarifying the installation steps, OS-compatibility, and community guidelines. I believe this submission is overall a high quality project, apart from a few minor clarifications that could be added to the readme. I love that the repository includes links to videos that show how the software works and what it does, which was helpful in determining that the software does work and accomplishes what the paper and readme declare.

I'm personally OK with having a non-standard license file name, though the license file is not complete and contains template placeholders. I created another issue thread for that.
