### 2022-01-28 7a137e0
Add installation guidelines
