### 2022-01-28 3ff3a0d
Refactor paper
