### 2022-01-30 #4061 @erik-whiting
> 👋 @erik-whiting, please update us on how your review is going (this is an automated reminder).

I opened two issues for the general checks and paper review. I have not had time to start the other reviews, but it's on my agenda, I promise 👍
