### 2022-01-31 ca094ad
Refactor paper
