### 2022-02-16 #4061 @AustinTSchaffer
@kasbot I looked through your `joss_review_round_1` branch, looks good! I'll close my issue threads on your repository once the branch is merged, and check off the rest of my checklist above.
