### 2022-02-20 SINTEF/AutoActive-ActivityPresenter#6 @erik-whiting
@AustinTSchaffer hey, I'm the other reviewer. I noticed they added Community Guidelines to their README file on the front page of the project. This was good enough for me, do you agree?
