### 2022-02-20 #4061 @erik-whiting
Update on my end: I think @kasbot 's changes are good so far, I will test out the functionality and note any performance improvements later this weekend hopefully. Thanks everyone for your patience so far.
