### 2022-02-20 SINTEF/AutoActive-ActivityPresenter#6 @AustinTSchaffer
Yes, thanks for the reminder.
