### 2022-03-05 ecd5011
Added Base implementations of IDataPoint and ITimePoint with viewers
