### 2022-03-05 54c594c
Added helper methods to search for typed children
