+++ UI/UI/XamarinHelpers.cs
@@ -78,6 +78,29 @@ namespace SINTEF.AutoActive.UI
             return list;
         }
 
+        public static T GetFirstChildElement<T>(Element element)
+        {
+            var elements = new Queue<Element>();
+            elements.Enqueue(element);
+
+            while (elements.Any())
+            {
+                var el = elements.Dequeue();
+
+                if (el is Layout layout)
+                {
+                    foreach (var child in layout.Children)
+                    {
+                        elements.Enqueue(child);
+                    }
+                }
+
+                if (el is T item)
+                    return item;
+            }
+            return default(T);
+        }
+
         public static async Task ShowOkMessage(string title, string message, Page page = null)
         {
             if (page == null)
