### 2022-03-05 6f36295
Simplified ArchiveStructure inheritance
