+++ Archive/Plugin/IArchivePlugin.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-
+using SINTEF.AutoActive.Databus.Interfaces;
 using SINTEF.AutoActive.Plugins;
 
 namespace SINTEF.AutoActive.Archive.Plugin
@@ -9,7 +9,7 @@ namespace SINTEF.AutoActive.Archive.Plugin
     [PluginType(AllowMultipleImplementations = false, UseSingletonInstance = true)]
     public interface IArchivePlugin
     {
-        Task<ArchiveStructure> CreateFromJSON(JObject json, Archive archive, Guid sessionId);
+        Task<IDataStructure> CreateFromJSON(JObject json, Archive archive, Guid sessionId);
     }
 
     public class ArchivePluginAttribute : PluginAttribute
