+++ Plugins.Import/Mqtt/TableTimeIndexDyn.cs
@@ -72,8 +72,7 @@ namespace SINTEF.AutoActive.Plugins.Import.Mqtt
         public ITimePoint TimePoint => time;
         public long Start => time.data[0];
         public long End => time.data[time.length - 1];
-
-        public int Length => time.length;
+        public long Start => time.Start;
 
         public event TimeViewerWasChangedHandler TimeChanged;
 
