+++ Databus/Implementations/TabularStructure/Columns/GenericColumn.cs
@@ -73,14 +73,12 @@ namespace SINTEF.AutoActive.Databus.Implementations.TabularStructure.Columns
                 throw new ArgumentException();
             if (Length <= 0) return new SpanPair<T1>();
 
-            var elements = _column.Data.AsSpan(StartIndex, Length);
-
             Span<T1> data;
             unsafe
             {
                 var mem = _column.Data.AsMemory(StartIndex, Length);
                 using (var pin = mem.Pin())
-                    data = new Span<T1>(pin.Pointer, elements.Length);
+                    data = new Span<T1>(pin.Pointer, Length);
             }
 
             return new SpanPair<T1>(StartIndex, Index.Data.AsSpan(StartIndex, Length), data);
