+++ UI/UI/Figures/LinePlot.cs
@@ -14,7 +14,6 @@ using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Linq;
 using SINTEF.AutoActive.UI.Interfaces;
 using Xamarin.Forms;
-using ITimeSeriesViewer = SINTEF.AutoActive.Databus.Common.ITimeSeriesViewer;
 
 namespace SINTEF.AutoActive.UI.Figures
 {
@@ -27,14 +26,13 @@ namespace SINTEF.AutoActive.UI.Figures
 
             var lineDrawer = await linePlot.CreateLineDrawer(datapoint);
             linePlot.AddLine(lineDrawer);
+
             return linePlot;
         }
 
         protected LinePlot(TimeSynchronizedContext context, IDataPoint dataPoint) : base(context, dataPoint)
         {
         }
-
-
 
         private double _previouseWindowHeight = 0;
         private double _previouseWindowWidth = 0;
@@ -48,9 +46,6 @@ namespace SINTEF.AutoActive.UI.Figures
             get => _axisValuesVisible && !_autoScaleIndependent;
             set => _axisValuesVisible = value;
         }
-
-
-
 
         protected override void RedrawCanvas(SKCanvas canvas, SKImageInfo info)
         {
