### 2022-03-05 c707454
Added events Start, End and DataChanged to IDataPoint
