### 2022-03-05 79e6ade
Added Changed event to IDataViewer
