### 2022-03-05 77a38e5
Added index-based color getter
