### 2022-03-05 3e08cec
Added support for visualizing base types
