+++ UI/UI/Views/FigureView.xaml.cs
@@ -21,10 +21,8 @@ using SINTEF.AutoActive.Databus.Implementations;
 
 namespace SINTEF.AutoActive.UI.Views
 {
-	public partial class FigureView : ContentView, ISerializableView
+    public partial class FigureView : ContentView, ISerializableView
     {
-        //public static Color ElementBackgroundColor = Color.CornflowerBlue.MultiplyAlpha(0.7d);
-
         public static Color ElementBackgroundColor = Color.FromHex("23A2B1");
         public List<IDataPoint> DataPoints { get; set; } = new List<IDataPoint>();
 
@@ -32,34 +30,34 @@ namespace SINTEF.AutoActive.UI.Views
         public string ViewType => StaticViewType;
 
         public TimeSynchronizedContext Context { get; }
-	    protected static readonly SKPaint FramePaint = new SKPaint
-	    {
-	        Color = SKColors.LightSlateGray,
-	        Style = SKPaintStyle.Stroke,
-	        StrokeWidth = 1,
-	        StrokeJoin = SKStrokeJoin.Miter,
-	        IsAntialias = false,
-	    };
+        protected static readonly SKPaint FramePaint = new SKPaint
+        {
+            Color = SKColors.LightSlateGray,
+            Style = SKPaintStyle.Stroke,
+            StrokeWidth = 1,
+            StrokeJoin = SKStrokeJoin.Miter,
+            IsAntialias = false,
+        };
 
-	    public bool Selected
-	    {
-	        get => SelectionFrame.BorderColor == Color.Red;
-	        set => SelectionFrame.BorderColor = value ? Color.Red : Color.Black;
-	    }
+        public bool Selected
+        {
+            get => SelectionFrame.BorderColor == Color.Red;
+            set => SelectionFrame.BorderColor = value ? Color.Red : Color.Black;
+        }
 
-	    public bool ContextButtonIsVisible
-	    {
-	        get => ContextButton.IsVisible;
-	        set => ContextButton.IsVisible = value;
-	    }
+        public bool ContextButtonIsVisible
+        {
+            get => ContextButton.IsVisible;
+            set => ContextButton.IsVisible = value;
+        }
 
         protected readonly SKPaint TextPaint = new SKPaint
-	    {
-	        Color = SKColors.Black,
-	        Style = SKPaintStyle.Fill,
-	        IsAntialias = true,
-	        SubpixelText = true,
-	    };
+        {
+            Color = SKColors.Black,
+            Style = SKPaintStyle.Fill,
+            IsAntialias = true,
+            SubpixelText = true,
+        };
 
         private readonly List<IDataViewer> _viewers = new List<IDataViewer>();
         protected List<IDataViewer> Viewers
@@ -157,16 +155,16 @@ namespace SINTEF.AutoActive.UI.Views
 
 
         protected virtual void RedrawCanvas(SKCanvas canvas, SKImageInfo info)
-	    {
-	        // Clear background and draw frame
-	        canvas.Clear(SKColors.White);
-	        canvas.DrawRect(0, 0, info.Width - 1, info.Height - 1, FramePaint);
+        {
+            // Clear background and draw frame
+            canvas.Clear(SKColors.White);
+            canvas.DrawRect(0, 0, info.Width - 1, info.Height - 1, FramePaint);
 
             const string text = "Override RedrawCanvas";
-	        var textHeight = TextPaint.FontMetrics.CapHeight;
-	        var maxTextWidth = TextPaint.MeasureText(text);
+            var textHeight = TextPaint.FontMetrics.CapHeight;
+            var maxTextWidth = TextPaint.MeasureText(text);
 
[…]
