### 2022-03-05 1cacd7b
Changed responsibility for drawing lines
