+++ UI/UI/InitializePlugins.cs
@@ -10,6 +10,7 @@ using SINTEF.AutoActive.Plugins.Import.Video;
 using SINTEF.AutoActive.Plugins.Import.Csv;
 using SINTEF.AutoActive.Plugins.Import.Gaitup;
 using SINTEF.AutoActive.Plugins.Import.Excel;
+using SINTEF.AutoActive.Plugins.Import.Json;
 using Xamarin.Forms;
 
 [assembly: Dependency(typeof(SINTEF.AutoActive.UI.InitializePlugins))]
@@ -23,6 +24,7 @@ namespace SINTEF.AutoActive.UI
             typeof(ArchiveSessionPlugin),
             typeof(ArchiveTablePlugin),
             typeof(ArchiveVideoPlugin),
+            typeof(AnnotationProvider),
 
             // Import plugins
             typeof(GarminImportPlugin),
@@ -32,7 +34,8 @@ namespace SINTEF.AutoActive.UI
             typeof(ImportGaitupPlugin),
             typeof(ImportCsvCatapult),
             typeof(ImportGenericExcel),
-            typeof(ImportGaitUpReults)
+            typeof(ImportGaitUpReults),
+            typeof(ImportAnnotationPlugin)
         };
     }
 }
