+++ Plugins.Import/Json/AnnotationParser.cs
@@ -1,0 +1,242 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using Newtonsoft.Json.Linq;
+using SINTEF.AutoActive.Databus.Implementations;
+using SINTEF.AutoActive.Databus.Interfaces;
+using SINTEF.AutoActive.FileSystem;
+using Newtonsoft.Json;
+using SINTEF.AutoActive.Archive.Plugin;
+using SINTEF.AutoActive.Archive;
+using ICSharpCode.SharpZipLib.Zip;
+
+namespace SINTEF.AutoActive.Plugins.Import.Json
+{
+    [ImportPlugin(".json")]
+    public class ImportAnnotationPlugin : IImportPlugin
+    {
+        public async Task<bool> CanParse(IReadSeekStreamFactory readerFactory)
+        {
+            var stream = await readerFactory.GetReadStream();
+            using (var streamReader = new StreamReader(stream))
+            using (var jsonReader = new JsonTextReader(streamReader))
+            {
+                var serializer = new JsonSerializer();
+                var json = (JToken)serializer.Deserialize(jsonReader);
+                if (json["AutoActiveType"].ToObject<string>() != "Annotation")
+                {
+                    return false;
+                }
+            }
+            return true;
+        }
+
+        public void GetExtraConfigurationParameters(Dictionary<string, (object, string)> parameters)
+        { }
+
+        public async Task<IDataProvider> Import(IReadSeekStreamFactory readerFactory,
+            Dictionary<string, object> parameters)
+        {
+            var importer = new AnnotationProvider(parameters, readerFactory.Name);
+            importer.ParseFile(await readerFactory.GetReadStream());
+            return importer;
+        }
+    }
+
+    [ArchivePlugin("no.sintef.annotation")]
+    public class AnnotationProvider : BaseDataProvider, IArchivePlugin, ISaveable
+    {
+        private AnnotationSet AnnotationSet
+        {
+            get => _annotationsBacking;
+            set
+            {
+                if(_annotationsBacking != null)
+                {
+                    RemoveDataPoint(DataPoint);
+                }
+                var annotationSet = value;
+                var annotations = annotationSet.Annotations;
+
+                _timeData = annotations.Select(annotation => annotation.Timestamp).ToList();
+                _annotationData = annotations.Select(annotation => annotation.Type).ToList();
+
+                IsSaved = false;
+
+                var isWorldSynchronized = annotationSet.IsWorldSynchronized;
+
+                _timePoint = new BaseTimePoint(_timeData, isWorldSynchronized);
+
+                var uri = "Annotations";
+                _dataPoint = new AnnotationDataPoint("Annotations", _annotationData, _timePoint, uri, null);
+
+                _annotationsBacking = annotationSet;
+                AddDataPoint(_dataPoint);
+            }
+        }
+        private AnnotationSet _annotationsBacking;
+
+        private List<long> _timeData;
+        private List<int> _annotationData;
+        private BaseTimePoint _timePoint;
+        private AnnotationDataPoint _dataPoint;
+
+        public AnnotationDataPoint DataPoint => _dataPoint;
+
+        public AnnotationProvider(Dictionary<string, object> parameters, string filename)
+        {
+            Name = parameters["Name"] as string;
+        }
+
+        public AnnotationProvider()
+        {
+            Name = "AnnotationProvider";
+        }
+
+        public static AnnotationProvider CreateNew()
+        {
+            var provider = new AnnotationProvider
+            {
+                AnnotationSet = new AnnotationSet()
+            };
+            return provider;
+        }
+
[…]
