### 2022-03-07 17f96b2
Added option to delete features
