### 2022-03-14 878511f
Added center line for annotations
