### 2022-03-17 6a908a3
Moved helper method for annotation provider
