### 2022-03-17 e7bec7b
Drawing annotation tags
