### 2022-03-17 5638905
Updated dependencies.
Xamarin 4.4.0 -> 5.0.0
CsvHelper 12.2.2 -> 27.2.1
Newtonsoft.Json 12.0.3 -> 13.0.1
Rg.Plugins.Popup 1.2 -> 2.0
SkiaSharp 1.68 -> 2.80
MetadataExtractor 2.2.0 -> 2.7.2
