### 2022-03-17 #4061 @prashjha
Hi @kasbot, see if these suggestions make sense for your draft:

- Line 24: should you use "optimization" here?
- Line 24: in the sentence "... and evaluation of algorithms to extract high-level, quantified analysis from sensor data", it seems there is something missing between "high-level" and "quantified" or maybe comma is inappropriate here?
- Line 27: sentence has many errors. E.g., it should be "consists". Further, you need to use commas and "and" to list three modules, for example, "ActivityPresenter, the accompanying Matlab, and Python toolboxes". Moreover, you say these three are open-source modules but Matlab is not open-source. Did you mean Matlab script that can be run using Octave? Also, "the accompanying Matlab" is meaningless. Did you mean to say the accompanying Matlab scripts? With little effort this sentence can be greatly improved.
- Line 64: "... sessions is a key to ..." and "store data in ARE; they are not ...".
- Line 67: "When a session is saved, it becomes ...".
- Line 68: "This allows sessions ...".

After you have updated the draft, can you also do (if not done already) a 'tagged' release of your code
archive the release using zenedo or other methods? Make sure that the title of zenedo archive matches with the title of this JOSS submission.

Once you are done, I will hand your paper to EiC for the final decision.
