### 2022-03-18 962649d
Transition towards DataBindings for AnnotationInfo
