### 2022-03-18 c11be72
Redrawing figures when exiting settings view
