### 2022-03-18 9727916
Handling drawing when Tag == null. Added text margin.
