### 2022-03-18 1d8d4b7
Added support for drawing annotations 10-39
