### 2022-03-22 1eeeb9f
Refactore paper
