### 2022-03-25 d16d0c5
Refactore paper based on comments from Anders and Torstein
