### 2022-03-29 3722c68
Fixed BaseTimePoint not updating correctly
