### 2022-03-29 36ef08d
Improved support for editing annotation info
