+++ UI/UI/Pages/Player/AnnotationsPopupView.xaml.cs
@@ -10,6 +10,9 @@ namespace SINTEF.AutoActive.UI.Pages.Player
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class AnnotationsPopupView : Rg.Plugins.Popup.Pages.PopupPage
     {
+        // The list is pre-populated with ID of existing annotations, plus 1 to [AnnotationPrepopulate]
+        private static readonly int AnnotationPrepopulate = 0;
+
         private readonly AnnotationSet _annotationSet;
         private int _rowCount = 0;
 
@@ -21,7 +24,14 @@ namespace SINTEF.AutoActive.UI.Pages.Player
             _annotationSet = provider.AnnotationSet;
             var annotationInfos = _annotationSet.AnnotationInfo;
 
-            for (var i = 1; i <= 20; i++)
+            foreach(var annotaion in _annotationSet.Annotations)
+            {
+                if (annotationInfos.ContainsKey(annotaion.Type)) continue;
+
+                annotationInfos[annotaion.Type] = new AnnotationInfo();
+            }
+
+            for (var i = 1; i <= AnnotationPrepopulate; i++)
             {
                 if (annotationInfos.ContainsKey(i)) continue;
 
@@ -32,6 +42,19 @@ namespace SINTEF.AutoActive.UI.Pages.Player
             {
                 AddLine(index, info);
             }
+
+            UpdateLineID();
+        }
+
+        private int NextLineID()
+        {
+            var existingIds = _annotationSet.AnnotationInfo.Keys;
+            return existingIds.Count == 0 ? 1 : existingIds.Max() + 1;
+        }
+
+        private void UpdateLineID()
+        {
+            LineID.Text = NextLineID().ToString();
         }
 
         private void AddLine(int index, AnnotationInfo info)
@@ -50,7 +73,7 @@ namespace SINTEF.AutoActive.UI.Pages.Player
             LayoutGrid.Children.Add(entryId);
 
             var propCount = 1;
-            foreach(var prop in new[] { "Name", "Tag", "Comment" })
+            foreach (var prop in new[] { "Name", "Tag", "Comment" })
             {
                 var entry = new Entry()
                 {
@@ -64,12 +87,34 @@ namespace SINTEF.AutoActive.UI.Pages.Player
             }
         }
 
-        private void AddLineButton_Clicked(object sender, EventArgs e)
+        private async void AddLineButton_Clicked(object sender, EventArgs e)
         {
-            var index = _annotationSet.AnnotationInfo.Keys.Max() + 1;
             var annotationInfo = new AnnotationInfo();
+            if (!int.TryParse(LineID.Text, out var index) || index < 0)
+            {
+                var indexSuggestion = NextLineID();
+                var res = await DisplayActionSheet($"Could not parse {LineID.Text} to a valid Annotation ID. Would you like to use {indexSuggestion}?", "Yes", "No", new string[] {});
+                switch(res)
+                {
+                    case "No":
+                        return;
+                    case "Yes":
+                        index = indexSuggestion;
+                        break;
+                }
+            }
+
+
+            if (_annotationSet.AnnotationInfo.ContainsKey(index))
+            {
+                await XamarinHelpers.ShowOkMessage("Already exists", $"Annotation Info with ID {index} already exists. Can not add duplicate.", this);
+                return;
+            }
+
             _annotationSet.AnnotationInfo[index] = annotationInfo;
             AddLine(index, annotationInfo);
+
+            UpdateLineID();
         }
 
         private async void CancelButton_Clicked(object sender, EventArgs e)
