### 2022-03-30 b10e688
Removed unused code
