### 2022-03-31 36f64cf
Fixed indexing when skipping data points.
