+++ UI/Tests/PlotTests.cs
@@ -11,7 +11,7 @@ namespace Tests
 {
     public class PlotTests
     {
-        private void TestDataSet(long[] timeArray, double[] dataArray, int maxNum)
+        private static ITimeSeriesViewer CreateTimeSeriesViewer(long[] timeArray, double[] dataArray)
         {
             Assert.Equal(timeArray.Length, dataArray.Length);
 
@@ -30,6 +30,12 @@ namespace Tests
             Assert.Equal(0, tsc.AvailableTimeFrom);
             Assert.Equal(timeArray.Last() - startTime, tsc.AvailableTimeTo);
             tsc.SetSelectedTimeRange(tsc.AvailableTimeFrom, tsc.AvailableTimeTo);
+            return viewer;
+        }
+
+        private void TestDataSet(long[] timeArray, double[] dataArray, int maxNum)
+        {
+            var viewer = CreateTimeSeriesViewer(timeArray, dataArray);
 
             var dataView = viewer.GetCurrentData<double>();
 
@@ -38,7 +44,7 @@ namespace Tests
             var count = 0;
             Assert.True(en.MoveNext());
             count++;
-            
+
             Assert.Equal((double)timeArray.First(), en.Current.x, 5);
             Assert.Equal(dataArray.First(), en.Current.y, 5);
             while (en.MoveNext())
@@ -83,6 +89,83 @@ namespace Tests
                 time[i] = 1000 * i + r.Next(0, 750);
             }
             TestDataSet(time, data, maxNum);
+        }
+
+        [Fact]
+        public void CheckAllData()
+        {
+            var timeArray = new long[] { 10, 20, 30, 40, 50, 60, 70, 80, 90, 100 };
+            var dataArray = new double[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+
+            var viewer = CreateTimeSeriesViewer(timeArray, dataArray) as TableColumnViewer;
+            viewer.SetTimeRange(timeArray.First(), timeArray.Last());
+            var data = viewer.GetCurrentData<double>();
+            var en = data.GetEnumerator(dataArray.Length*10);
+
+            var index = 0;
+            while(en.MoveNext())
+            {
+                Assert.Equal(timeArray[index], en.Current.x);
+                Assert.Equal(dataArray[index], en.Current.y);
+                index++;
+            }
+        }
+
+        [Fact]
+        public void TestCurrentDataSelector()
+        {
+            var timeArray = new long[] { 0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100 };
+            var dataArray = new double[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+
+            var viewer = CreateTimeSeriesViewer(timeArray, dataArray) as TableColumnViewer;
+            viewer.SetTimeRange(20, 40);
+            //viewer.StartIndex = 1;
+            var data = viewer.GetCurrentData<double>();
+            var en = data.GetEnumerator(100);
+
+            Assert.True(en.MoveNext());
+            Assert.Equal(20, en.Current.x);
+            Assert.True(en.MoveNext());
+            Assert.Equal(30, en.Current.x);
+            Assert.True(en.MoveNext());
+            Assert.Equal(40, en.Current.x);
+            //Assert.False(en.MoveNext());
+        }
+
+        [Fact]
+        public void ModuloEnumeratorTest()
+        {
+            var size = 1000;
+
+            var dataArray = new double[size];
+            var timeArray = new long[size];
+            for (var i = 0; i < size; i++)
+            {
+                timeArray[i] = i;
+                dataArray[i] = i;
+            }
+
+            var viewer = CreateTimeSeriesViewer(timeArray, dataArray) as TableColumnViewer;
[…]
