### 2022-03-31 b4e408e
Made sorted time seach more robust
