### 2022-03-31 #4061 @kasbot
Hi @prashjha,
Thank you for your suggestions. The draft has now been updated according to your suggestions. We have also done a tagged release of our code using zenodo which can be found here https://zenodo.org/search?page=1&size=20&q=AutoActive
