### 2022-04-18 #4061 @arfon
@kasbot – I made a few minor edits to the paper here: https://github.com/SINTEF/AutoActive-ActivityPresenter/pull/15
