### 2022-04-21 #4061 @aliverud
Thanks a lot for the review and publishing!

There is a broken link for the author "Anders E. Liverud" at the publication page. Could you please correct? (the ".E" is missing in the authors page: https://joss.theoj.org/papers/by/Anders%20Liverud vs. https://joss.theoj.org/papers/by/Anders%20E.%20Liverud)
